reformatted: token serializer to serializer file
